Refactor account deletion from serverside with error handling

+++ server/src/controllers/accountDelete.js
@@ -1,21 +1,30 @@
-const {supabase, supabaseAdmin} = require('../config/supabase')
-
-async function rollbackAccountCreation(userId) {
-    // Deletes the user from the Auth Database
-      await supabaseAdmin
-        .auth
-        .admin
-        .deleteUser(userId)
-      
-      // Deletes the user from the Account Database
-      await supabaseAdmin
-        .from('account')
-        .delete()
-        .eq('id', userId)
-}
+const { supabaseAdmin } = require('../config/supabase');
 
 async function deleteAccount(userId) {
+  if (!userId) {
+    throw new Error('User ID is required');
+  }
 
+  const { error: authError } = await supabaseAdmin
+                                    .auth
+                                    .admin
+                                    .deleteUser(userId);
+  if (authError) {
+    throw authError;
+  }
+
+  const { error: accountError } = await supabaseAdmin
+    .from('account')
+    .delete()
+    .eq('id', userId);
+
+  if (accountError) {
+    throw accountError;
+  }
 }
 
-module.exports = { rollbackAccountCreation, deleteAccount }
+async function rollbackAccountCreation(userId) {
+  await deleteAccount(userId);
+}
+
+module.exports = { rollbackAccountCreation, deleteAccount };
